rbs/requirements.txt
# Copyright 2018 The Bazel Authors.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# For the version of boto3 and botocore, see
# https://docs.aws.amazon.com/lambda/latest/dg/current-supported-versions.html
boto3==1.7.0
botocore==1.9.3
requests==2.18.4
enum34==1.1.6
six==1.11.0
mock==1.3.0
setuptools==39.0.1
requests-mock==1.5.0
attrs==18.1.0
requests-aws4auth==0.9
jsonschema==2.6.0
PyYAML==3.12
pyOpenSSL==18.0.0
cryptography===2.2.2
idna==2.6
certifi==2018.4.16
